backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
pandas==2.2.2
requests==2.31.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
pydantic==2.7.1
pydantic-settings==2.2.1
aiofiles==23.2.1

# Scraper dependencies
feedparser==6.0.12
textblob==0.19.0
pytrends==4.9.2
google-search-results==2.4.2
deep-translator==1.11.4
newspaper3k==0.2.8
nltk==3.9.2
playwright==1.58.0
